# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DroniAuth API",
        "description": "OAuth 2.0 Authorization Server. Provides authentication, user management, and OAuth client management endpoints.",
        "contact": {
            "email": "dev@droni.co"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://auth.droni.co",
            "description": "Production"
        },
        {
            "url": "http://localhost",
            "description": "Local development"
        }
    ],
    "paths": {
        "/login": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Show login page",
                "operationId": "d38b0625602ec8e41c3bb98aa2e969de",
                "responses": {
                    "200": {
                        "description": "Login page (HTML/Inertia)"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and create session",
                "operationId": "42f77e072dec84b0e1094cca1228298e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirects to dashboard or two-factor challenge if 2FA is enabled"
                    },
                    "422": {
                        "description": "Invalid credentials or validation error"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Invalidate session and log out",
                "operationId": "03afb12b342c2ea43bdc1de4f39e91a8",
                "responses": {
                    "302": {
                        "description": "Redirects to /"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/reset-password/{token}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Show password reset form",
                "operationId": "139cbd7a33506f23db79a42dd3a1b825",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset form (HTML/Inertia)"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset user password using token from email",
                "operationId": "24b1c68468bf984b8806908f211f3827",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Reset token from email link",
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Password reset – redirects to /login. All tokens for this user are revoked."
                    },
                    "422": {
                        "description": "Invalid token, expired link, or weak password"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "description": "Session cookie (web authentication)",
                "name": "droniauth_session",
                "in": "cookie"
            },
            "oauth2": {
                "type": "oauth2",
                "description": "OAuth 2.0 Authorization Code flow issued by this server",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": {
                            "profile": "Read user profile (name, avatar)",
                            "email": "Read email address",
                            "roles": "Read roles and permissions",
                            "admin": "Full admin access"
                        }
                    }
                }
            }
        }
    },
...
}